// src/enums/logAction.ts
export enum LogAction {
  init = 'Init',
  initPackages = 'InitPackages',
  connect = 'ConnectCoreAcc',
  disconnect = 'DisconnectCoreAcc',
  forceDisconnect = 'ForceDisconnectCoreAcc',
  forceUnpublish = 'ForceUnpublishCoreAcc',
  getAccPack = 'GetAccPack',
  signal = 'SignalCoreAcc',
  startCall = 'StartCallCoreAcc',
  endCall = 'EndCallCoreAcc',
  toggleLocalAudio = 'ToggleLocalAudio',
  toggleLocalVideo = 'ToggleLocalVideo',
  toggleRemoteAudio = 'ToggleRemoteAudio',
  toggleRemoteVideo = 'ToggleRemoteVideo',
  subscribe = 'SubscribeCoreAcc',
  unsubscribe = 'UnsubscribeCoreAcc'
}